feat(NavigationItem.tsx): add NavigationItem component
This commit adds a new component called NavigationItem. It is a functional component that takes a NavItem object as a prop and renders a link to the path specified in the NavItem object. The component also checks if the current pathname matches the NavItem path and applies a border and text color to indicate that the link is active. The component uses the Finlandica font from Google Fonts and the usePathname hook from the next/navigation package.